refactor get_pending_txn
